Cleaned up code

+++ js/game.js
@@ -1,22 +1,3 @@
-// let player
-// let bg
-// let rec
-// let cursors
-// let weapon
-// let fireButton
-// let enemy
-// let tween
-// let boundary
-// let walk
-// let boundaries
-// let teleport
-// let teleport2
-// let log
-// let levelUnlock
-// let facing
-// let enemyWeapon
-// let waterBoundary
-
 let Game = {
   preload: function() {
     game.load.image('lvl1bg', 'assets/images/lvl1bg.png')
@@ -37,24 +18,9 @@ let Game = {
     // Add Background
     bg = makeSprite(0, 0, 'lvl1bg')
 
-    // Add log, has to be called before player so that
-    // player appears on top
-    
-    // Add Player
-    // player = game.add.sprite(240, 1304, 'hamster')
-    
-    // Player locations for testing
-    // player = game.add.sprite(700, 700, 'hamster')
-    // log = game.add.sprite(370, 200, 'log')
-
+    // Add Sprites
     log = makeSprite(370, 200, 'log')
-    player = makeSprite(115, 460, 'hamster')
-    enemy = makeSprite(150, 500, 'hilary2')
-
-    // Create
-    // enemies = game.add.group()
-    // // enemies.create(150,500,'hilary2')
-    // enemies.create(150,700,'hilary2')
+    player = makeSprite(240, 1304, 'hamster')
     
     // Animations  
     initPlayerAnimations(player)
@@ -66,20 +32,6 @@ let Game = {
 
     // Weapon Methods   
     initWeapon(weapon)
-    
-    // ENEMY WEAPON //
-    
-    // Add  Enemy Weapon
-    enemyWeapon = game.add.weapon(30, 'shuriken')
-    
-    // Weapon Methods
-    initEnemyWeapon(enemyWeapon)
-    enemyWeapon.fireAngle = Phaser.ANGLE_DOWN
-    enemyWeapon.trackSprite(enemy, 100, 100, false)
-    
-    // Add Enemy
-    // enemy = game.add.sprite(400, 200, 'hilary2')
-    tween = game.add.tween(enemy)
     
     boundaries = game.add.group()
     boundaries.add(makeWorldSprite(0, 335, 670, 90, 'bg'))
@@ -111,21 +63,19 @@ let Game = {
     logCheck = makeSprite(370, 450, 'log')
     alpha(logCheck)
 
+    someLog = makeSprite(370, 450, 'log')
+    alpha(someLog)
+
     // Treasure
     treasure = makeSprite(840, 670, 'treasure')
 
     // Level Unlock
     levelUnlock = makeSprite(640, 800, 'log')
 
-    // Moving enemy
-    tween.to({ x: 700 }, 4000, 'Linear', true, 0, 20, true).loop(true)
-    
     // Enable physics   
     game.physics.enable([
       player, 
       weapon,
-      enemyWeapon, 
-      // enemy, 
       teleport, 
       teleport2, 
       boundaries, 
@@ -148,10 +98,6 @@ let Game = {
     // Water Death
     collisionGroup(waterBoundaries)
 
-    // Log Check
-    collideImmovable(log)
-    collideImmovable(logCheck)
-
     // World Boundaries
     collisionGroup(boundaries)
 
@@ -170,10 +116,8 @@ let Game = {
   
   update: function() {
     // game.physics.arcade.collide(player, rec, this.startLevelTwo, null, this)
-    // game.physics.arcade.collide(player, enemy, this.killPlayer, null, this)
     game.physics.arcade.collide(player, log, this.moveLog, null, this)
     game.physics.arcade.collide(player, waterBoundaries, this.killPlayer, null, this)
-    // game.physics.arcade.overlap(weapon.bullets, enemy, this.killEnemy, null, this)
     game.physics.arcade.collide(player, teleport, this.teleportPlayer, null, this)
     game.physics.arcade.collide(player, teleport2, this.teleportPlayer2, null, this)
     game.physics.arcade.collide(player, boundaries)
@@ -215,21 +159,29 @@ let Game = {
 
   // Move the log
   moveLog: function (player, logs) {
+    game.world.bringToTop(player)
+    
     this.logMoved = true
     
     if (this.logMoved) {
-      logCheck.kill()
-      timeDelay(1000, logs, 370, 450)
+      game.time.events.add(500, () => {
+        logs.reset(logs.body.x = 370, logs.body.y = 450)
+        logs.kill()
+      })
+
+      game.time.events.add(500, () => {
+        someLog.alpha = true
+      })
     }
   },
 
-  // Check if log has been moved
   checkPlatfrom: function(player, logCheck) {
-    player.kill()
-    timeDelay(500, player, 240, 1304)
+    if (!this.logMoved) {
+      player.kill()
+      timeDelay(500, player, 240, 1304)
+    }
   },
 
-  // Spawn Weapon
   spawnWeapon: function(player, treasure) {
     this.enableWeapon = true
     treasure.kill()
